requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
aiosqlite==0.21.0
altgraph==0.17.4
anyio==4.9.0
APScheduler==3.11.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==24.3.0
cattrs==24.1.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
exceptiongroup==1.2.2
frozenlist==1.5.0
graphviz==0.20.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.0.0
isodate==0.7.2
Jinja2==3.1.5
lxml==5.3.0
macholib==1.16.3
MarkupSafe==3.0.2
multidict==6.2.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
propcache==0.3.1
psycopg2-binary==2.9.10
pycparser==2.22
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
pytest==8.3.4
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-cache==1.2.1
requests-file==2.1.0
requests-toolbelt==1.0.0
responses==0.25.6
six==1.17.0
sniffio==1.3.1
tomli==2.2.1
typing_extensions==4.12.2
tzlocal==5.2
url-normalize==1.4.3
urllib3==2.3.0
yamllint==1.35.1
yarl==1.18.3
zeep==4.3.1
zipp==3.21.0
